Correção do EDITAR Aluno, implantação para gerar Excel e PDF

+++ database/migrations/2024_08_05_145409_create_alunos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('cpf');
             $table->string('telefone');
             $table->string('whatsapp');
-            $table->integer('curso_id')->nullable();
+            $table->unsignedBigInteger('curso_id')->nullable();
             $table->foreign('curso_id')->references('id')->on('cursos')->onDelete('cascade');
             $table->timestamps();
             $table->softDeletes();
